# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Category = typeof Category[keyof typeof Category];


export const Category = {
  Electronics: 'Electronics',
  Clothing: 'Clothing',
  'IDs_&_Wallets': 'IDs & Wallets',
  'Books_&_Stationery': 'Books & Stationery',
  Keys: 'Keys',
  Accessories: 'Accessories',
  Miscellaneous: 'Miscellaneous',
} as const;

export type ItemStatus = typeof ItemStatus[keyof typeof ItemStatus];


export const ItemStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export type ItemReportType = typeof ItemReportType[keyof typeof ItemReportType];


export const ItemReportType = {
  lost: 'lost',
  found: 'found',
} as const;

export interface Item {
  id: string;
  report_type: ItemReportType;
  category: Category;
  title: string;
  description: string;
  image_url: string;
  location: string;
  event_time: string;
  contact_info: string;
  status: ItemStatus;
  created_at: string;
}

export type ItemInputReportType = typeof ItemInputReportType[keyof typeof ItemInputReportType];


export const ItemInputReportType = {
  lost: 'lost',
  found: 'found',
} as const;

export interface ItemInput {
  report_type: ItemInputReportType;
  category: Category;
  /**
     * @minLength 3
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  description: string;
  image_url: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  location: string;
  event_time: string;
  /**
     * @minLength 3
     * @maxLength 255
     */
  contact_info: string;
}

export interface ItemUpdate {
  status: ItemStatus;
}

export interface ItemMatch {
  id: string;
  lost_item_id: string;
  found_item_id: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence_score: number;
  explanation: string;
  created_at: string;
  matched_item: Item;
}

export interface DashboardStats {
  active_lost: number;
  active_found: number;
  resolved: number;
  total_reports: number;
}

export interface Error {
  error: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Report not found
 */
export type NotFoundResponse = Error;

/**
 * Server error
 */
export type ServerErrorResponse = Error;

export type ReportTypeParameter = typeof ReportTypeParameter[keyof typeof ReportTypeParameter];


export const ReportTypeParameter = {
  lost: 'lost',
  found: 'found',
} as const;

export type CategoryParameter = Category;

export type ListItemsParams = {
type?: ReportTypeParameter;
category?: CategoryParameter;
/**
 * @maxLength 100
 */
search?: string;
date_from?: string;
date_to?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type GetRecentItemsParams = {
/**
 * @minimum 1
 * @maximum 12
 */
limit?: number;
};