# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */

export type Category = typeof Category[keyof typeof Category];


export const Category = {
  Electronics: 'Electronics',
  Clothing: 'Clothing',
  'IDs_&_Wallets': 'IDs & Wallets',
  'Books_&_Stationery': 'Books & Stationery',
  Keys: 'Keys',
  Accessories: 'Accessories',
  Miscellaneous: 'Miscellaneous',
} as const;