# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  active_lost: number;
  active_found: number;
  resolved: number;
  total_reports: number;
}