# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { ItemReportType } from './itemReportType';
import type { ItemStatus } from './itemStatus';

export interface Item {
  id: string;
  report_type: ItemReportType;
  category: Category;
  title: string;
  description: string;
  image_url: string;
  location: string;
  event_time: Date;
  contact_info: string;
  status: ItemStatus;
  created_at: Date;
}