# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { ItemInputReportType } from './itemInputReportType';

export interface ItemInput {
  report_type: ItemInputReportType;
  category: Category;
  /**
     * @minLength 3
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  description: string;
  image_url: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  location: string;
  event_time: Date;
  /**
     * @minLength 3
     * @maxLength 255
     */
  contact_info: string;
}