# lib/api-zod/src/generated/types/itemMatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { Item } from './item';

export interface ItemMatch {
  id: string;
  lost_item_id: string;
  found_item_id: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence_score: number;
  explanation: string;
  created_at: Date;
  matched_item: Item;
}