# lib/api-zod/src/generated/types/itemStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */

export type ItemStatus = typeof ItemStatus[keyof typeof ItemStatus];


export const ItemStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;