# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Lost & Found API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemStatus } from './itemStatus';

export interface ItemUpdate {
  status: ItemStatus;
}